+++ source/scriptlib/Lexer.h
@@ -44,7 +44,7 @@ typedef struct Token {
 	MY_TOKEN_TYPE theType;
 	CHAR theSource[MAX_TOKEN_LENGTH + 1];
 	TEXTPOS theTextPosition;
-	ULONG charOffset;
+	u32 charOffset;
 } Token;
 
 /******************************************************************************
@@ -63,7 +63,7 @@ typedef struct Lexer {
 
 
 //Constructor
-void Token_Init(Token * ptoken, MY_TOKEN_TYPE theType, const char* theSource, TEXTPOS theTextPosition, ULONG charOffset);
+void Token_Init(Token * ptoken, MY_TOKEN_TYPE theType, const char* theSource, TEXTPOS theTextPosition, u32 charOffset);
 void Lexer_Init(Lexer * plexer, pp_context * pcontext, const char* thePath, char* theSource, TEXTPOS theStartingPosition);
 void Lexer_Clear(Lexer * plexer);
 s32 Lexer_GetNextToken(Lexer * plexer, Token * theNextToken);
